create new library due to circular import dependencies
